YearPickerDialog click event
