+++ app/src/main/java/com/jp_ais_training/keibo/activity/MainActivity.kt
@@ -30,7 +30,6 @@ class MainActivity : AppCompatActivity() {
                 R.id.circle_statistics -> setCurrentFragment(secondFragment)
                 R.id.bar_statistics -> setCurrentFragment(thirdFragment)
                 R.id.settings -> setCurrentFragment(fourthFragment)
-
             }
             true
         }
