+++ app/src/main/java/com/jp_ais_training/keibo/frament/ContentsFragment.kt
@@ -450,6 +450,7 @@ class ContentsFragment() : Fragment() {
                     )
                 }
 
+                //$$
                 categoryDialog.setOnClickedListener(object : CategoryDialog.ButtonClickListener {
                     override fun onClicked(id: Int, name: String, type: String) {
                         if (type == "main") {
