+++ app/src/main/res/layout/fragment_bar_statistics.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -9,114 +8,29 @@
     <ScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:ignore="SpeakableTextPresentCheck">
-
+       >
         <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
             android:orientation="vertical">
 
             <FrameLayout
+                android:id="@+id/fl_year"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent">
+                android:layout_height="wrap_content"
+                />
 
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:orientation="vertical">
+            <FrameLayout
+                android:id="@+id/fl_month"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"  />
 
-                    <LinearLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:orientation="vertical">
-
-                        <LinearLayout
-                            android:layout_width="match_parent"
-                            android:layout_height="match_parent"
-                            android:background="@color/blue"
-                            android:orientation="vertical">
-
-                            <TextView
-                                android:id="@+id/textView4"
-                                android:layout_width="match_parent"
-                                android:layout_height="35dp"
-                                android:gravity="center_vertical|center_horizontal"
-                                android:text="年間変化統計"
-                                android:textSize="20dp"
-                                android:textColor="	#FFFFFF"
-                                android:layout_marginTop="15dp"/>
-
-                            <LinearLayout
-                                android:layout_width="match_parent"
-                                android:layout_height="wrap_content"
-                                android:gravity="center"
-                                android:orientation="horizontal">
-
-                                <ImageButton
-                                    android:id="@+id/date_left_btn"
-                                    android:layout_width="48dp"
-                                    android:layout_height="match_parent"
-                                    android:background="@drawable/navi1_foreground" />
-
-                                <Button
-                                    android:id="@+id/date_year_btn"
-                                    android:layout_width="wrap_content"
-                                    android:layout_height="wrap_content"
-                                    android:text="2022"
-                                    android:textColor="	#FFFFFF" />
-
-                                <ImageButton
-                                    android:id="@+id/date_right_btn"
-                                    android:layout_width="48dp"
-                                    android:layout_height="match_parent"
-                                    android:background="@drawable/navi2_foreground" />
-
-                            </LinearLayout>
-                        </LinearLayout>
-
-                        <LinearLayout
-                            android:layout_width="match_parent"
-                            android:layout_height="465dp"
-                            android:background="@color/yellow"
-                            android:orientation="vertical">
-
-                            <LinearLayout
-                                android:layout_width="match_parent"
-                                android:layout_height="wrap_content"
-                                android:gravity="end"
-                                android:orientation="horizontal">
-
-                                <Button
-                                    android:id="@+id/ei_button"
-                                    android:layout_width="70dp"
-                                    android:layout_height="wrap_content"
-                                    android:layout_gravity="center"
-                                    android:text="支出"
-                                    android:textColor="	#FFFFFF" />
-
-                                <Button
-                                    android:id="@+id/ii_button"
-                                    android:layout_width="70dp"
-                                    android:layout_height="wrap_content"
-                                    android:layout_gravity="center"
-                                    android:layout_marginRight="15dp"
-                                    android:text="収入" />
-
-                            </LinearLayout>
-
-                            <com.github.mikephil.charting.charts.BarChart
-                                android:id="@+id/barchart"
-                                android:layout_width="match_parent"
-                                android:layout_height="381dp"
-                                android:layout_marginTop="10dp" />
-
-                        </LinearLayout>
-                    </LinearLayout>
-
-                </LinearLayout>
-            </FrameLayout>
+            <FrameLayout
+                android:id="@+id/fl_week"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"  />
 
         </LinearLayout>
     </ScrollView>
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
